Add indicators provider
